+++ frontend/benefit/handler/src/components/applicationReview/salaryBenefitCalculatorView/SalaryBenefitCalculatorView.tsx
@@ -1,10 +1,5 @@
 import ReviewSection from 'benefit/handler/components/reviewSection/ReviewSection';
-import {
-  CALCULATION_DESCRIPTION_ROW_TYPES,
-  CALCULATION_SUMMARY_ROW_TYPES,
-  CALCULATION_TOTAL_ROW_TYPE,
-  CALCULATION_TYPES,
-} from 'benefit/handler/constants';
+import { CALCULATION_TYPES } from 'benefit/handler/constants';
 import { useCalculatorData } from 'benefit/handler/hooks/useCalculatorData';
 import {
   PaySubsidy,
@@ -39,6 +34,7 @@ import {
   $CalculatorText,
 } from '../ApplicationReview.sc';
 import CalculatorErrors from '../calculatorErrors/CalculatorErrors';
+import SalaryCalculatorTable from './EmploymentCalculatorTable/SalaryCalculatorTable';
 import SalaryBenefitManualCalculatorView from './SalaryBenefitManualCalculatorView';
 import { useSalaryBenefitCalculatorData } from './useSalaryBenefitCalculatorData';
 
@@ -594,34 +590,7 @@ const SalaryBenefitCalculatorView: React.FC<
           </$Notification>
         </$GridCell>
       )}
-
-      <$GridCell $colSpan={7}>
-        {data?.calculation?.rows &&
-          data?.calculation?.rows.map((row) => {
-            const isSummaryRowType = CALCULATION_SUMMARY_ROW_TYPES.includes(
-              row.rowType
-            );
-            const isTotalRowType = CALCULATION_TOTAL_ROW_TYPE === row.rowType;
-            const isDescriptionRowType =
-              CALCULATION_DESCRIPTION_ROW_TYPES.includes(row.rowType);
-            return (
-              <div key={row.id}>
-                <$CalculatorTableRow isTotal={isSummaryRowType}>
-                  <$ViewField isBold={isTotalRowType || isDescriptionRowType}>
-                    {row.descriptionFi}
-                  </$ViewField>
-                  {!isDescriptionRowType && (
-                    <$ViewField isBold={isTotalRowType}>
-                      {t(`${translationsBase}.tableRowValue`, {
-                        amount: formatStringFloatValue(row.amount),
-                      })}
-                    </$ViewField>
-                  )}
-                </$CalculatorTableRow>
-              </div>
-            );
-          })}
-      </$GridCell>
+      <SalaryCalculatorTable data={data} />
     </ReviewSection>
   );
 };
